+++ resources/recipes/configure.rb
@@ -34,6 +34,15 @@ rb_selinux_config 'Configure Selinux' do
   end
 end
 
+# Sudoers
+template '/etc/sudoers.d/redborder-manager' do
+  source 'redborder-manager.erb'
+  owner 'root'
+  group 'root'
+  mode '0440'
+  retries 2
+end
+
 consul_config 'Configure Consul Server' do
   cdomain node['redborder']['cdomain']
   dns_local_ip node['consul']['dns_local_ip']
@@ -195,6 +204,7 @@ druid_historical 'Configure Druid Historical' do
     name node['hostname']
     ipaddress node['ipaddress_sync']
     memory_kb node['redborder']['memory_services']['druid-historical']['memory']
+    maxsize node['redborder']['manager']['hd_services_current']['druid-historical'].to_i
     action [:add, :register]
   else
     action [:remove, :deregister]
@@ -491,7 +501,7 @@ rbale_config 'Configure redborder-ale' do
 end
 
 rblogstatter_config 'Configure redborder-logstatter' do
-  if manager_services['rb-logstatter']
+  if manager_services['rb-logstatter'] && manager_services['logstash'] && node.run_state['pipelines'] && !node.run_state['pipelines'].empty?
     action :add
   else
     action :remove
@@ -665,15 +675,6 @@ unless ssh_secrets.empty?
     retries 2
     variables(public_rsa: ssh_secrets['public_rsa'])
   end
-end
-
-# Sudoers
-template '/etc/sudoers.d/redborder-manager' do
-  source 'redborder-manager.erb'
-  owner 'root'
-  group 'root'
-  mode '0440'
-  retries 2
 end
 
 # Pending Changes..